requirements.txt
ultralytics
opencv-python
PyYAML
gpsd-py3
